Refs #124. Moved pagebreak testing to its own module
