+++ tests/export/html/test_page_break.py
@@ -1,0 +1,80 @@
+# coding: utf-8
+
+from __future__ import (
+    absolute_import,
+    print_function,
+    unicode_literals,
+)
+
+from pydocx.test import DocumentGeneratorTestCase
+from pydocx.test.utils import WordprocessingDocumentFactory
+from pydocx.wordml import MainDocumentPart
+
+
+class PageBreakTestCase(DocumentGeneratorTestCase):
+    def test_before_text_run(self):
+        document_xml = '''
+            <p>
+              <r>
+                <t>aaa</t>
+              </r>
+            </p>
+            <p>
+              <r>
+                <br type="page"/>
+                <t>bbb</t>
+              </r>
+            </p>
+        '''
+
+        document = WordprocessingDocumentFactory()
+        document.add(MainDocumentPart, document_xml)
+
+        expected_html = '<p>aaa</p><p><hr />bbb</p>'
+        self.assert_document_generates_html(document, expected_html)
+
+    def test_between_paragraphs(self):
+        document_xml = '''
+            <p>
+              <r>
+                <t>aaa</t>
+              </r>
+            </p>
+            <p>
+              <r>
+                <br type="page"/>
+              </r>
+            </p>
+            <p>
+              <r>
+                <t>bbb</t>
+              </r>
+            </p>
+        '''
+
+        document = WordprocessingDocumentFactory()
+        document.add(MainDocumentPart, document_xml)
+
+        expected_html = '<p>aaa</p><p><hr /></p><p>bbb</p>'
+        self.assert_document_generates_html(document, expected_html)
+
+    def test_after_text_run(self):
+        document_xml = '''
+            <p>
+              <r>
+                <t>aaa</t>
+                <br type="page"/>
+              </r>
+            </p>
+            <p>
+              <r>
+                <t>bbb</t>
+              </r>
+            </p>
+        '''
+
+        document = WordprocessingDocumentFactory()
+        document.add(MainDocumentPart, document_xml)
+
+        expected_html = '<p>aaa<hr /></p><p>bbb</p>'
+        self.assert_document_generates_html(document, expected_html)
