+++ tests/export/test_html.py
@@ -310,72 +310,3 @@ class FakedSuperScriptTestCase(DocumentGeneratorTestCase):
 
         expected_html = '<p>nth</p>'
         self.assert_document_generates_html(document, expected_html)
-
-
-class PageBreakTestCase(DocumentGeneratorTestCase):
-    def test_before_text_run(self):
-        document_xml = '''
-            <p>
-              <r>
-                <t>aaa</t>
-              </r>
-            </p>
-            <p>
-              <r>
-                <br type="page"/>
-                <t>bbb</t>
-              </r>
-            </p>
-        '''
-
-        document = WordprocessingDocumentFactory()
-        document.add(MainDocumentPart, document_xml)
-
-        expected_html = '<p>aaa</p><p><hr />bbb</p>'
-        self.assert_document_generates_html(document, expected_html)
-
-    def test_between_paragraphs(self):
-        document_xml = '''
-            <p>
-              <r>
-                <t>aaa</t>
-              </r>
-            </p>
-            <p>
-              <r>
-                <br type="page"/>
-              </r>
-            </p>
-            <p>
-              <r>
-                <t>bbb</t>
-              </r>
-            </p>
-        '''
-
-        document = WordprocessingDocumentFactory()
-        document.add(MainDocumentPart, document_xml)
-
-        expected_html = '<p>aaa</p><p><hr /></p><p>bbb</p>'
-        self.assert_document_generates_html(document, expected_html)
-
-    def test_after_text_run(self):
-        document_xml = '''
-            <p>
-              <r>
-                <t>aaa</t>
-                <br type="page"/>
-              </r>
-            </p>
-            <p>
-              <r>
-                <t>bbb</t>
-              </r>
-            </p>
-        '''
-
-        document = WordprocessingDocumentFactory()
-        document.add(MainDocumentPart, document_xml)
-
-        expected_html = '<p>aaa<hr /></p><p>bbb</p>'
-        self.assert_document_generates_html(document, expected_html)
